+++ src/Model/AbstractManager.php
@@ -30,7 +30,7 @@ abstract class AbstractManager
             $query .= ' ORDER BY ' . $orderBy . ' ' . $direction;
         }
 
-        return $this->pdo->query($query)->fetchAll();
+        return $this->pdo->query($query)->fetchObject();
     }
 
     /**
